Computer Vision Camera calibration
Below is the steps for implemented it.
1. Mark the work plane coordination for the corner on the cube
and provide the input box to key the corner coordinate value
before click the corresponding point on image.After that this
program will save the world plane coordinates and image plane
coordinates value to files.
2.find out M*, T*, R*, U0, V0
3. Ransac algorism

Below is the step for Ransac algorism.
1. A sample subset containing minimal data items is randomly
selected from the points. A fitting model and the corresponding
model parameter are computed using the small subset.
2. Algorithm check which elements of the entire points are
inlier and using the inlier points to generate a fit model and
calculate the corresponding model parameters using the inlier points.
3. Ransac algorithm will repeat the above two steps until k > maxk
or find out the best solution.